TRIX-261 roll back on invalid interface

+++ daemon/base/interface.py
@@ -182,15 +182,14 @@ class Interface():
                         )
                     elif (not macaddress) and (not options):
                         # this means we just made an empty interface. a no no - Antoine
-                        if not existing:
-                            self.delete_node_interface(nodeid=nodeid, interface=interface_name)
                         result=False
                         message="Invalid request: missing minimal parameters"
                     elif not existing:
-                        if not existing:
-                            self.delete_node_interface(nodeid=nodeid, interface=interface_name)
                         result=False
                         message="Invalid request: missing minimal parameters"
+                    if result is False and not existing:
+                        # roll back
+                        self.delete_node_interface(nodeid=nodeid, interface=interface_name)
 
                 if result is False:
                     response = f'{message}'
